Move locations to utils package
